feat(cld/runtime): add support to execute registered changesets from YAML input
Add shared durable pipeline YAML parsing helpers and introduce runtime execution for registered changesets from YAML in order. This unifies CLI/runtime input handling and lets downstream integration tests consume the same YAML-driven execution path.
