+++ .changeset/smart-files-brake.md
@@ -1,0 +1,5 @@
+---
+"chainlink-deployments-framework": minor
+---
+
+feat(cld/runtime): add support to execute registered changesets from YAML input
